python-rpc-service/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
python-dateutil>=2.8.0
pydantic>=1.10.0
loguru>=0.6.0
uvicorn>=0.20.0
fastapi>=0.85.0
httpx>=0.24.0
